+++ plugins/flutter_singbox_vpn/android/src/main/kotlin/com/tecclub/flutter_singbox/bg/DefaultNetworkMonitor.kt
@@ -46,6 +46,8 @@ object DefaultNetworkMonitor {
 
     suspend fun stop() {
         android.util.Log.d("DefaultNetworkMonitor", "Stopping network monitor")
+        listener = null
+        defaultNetwork = null
         DefaultNetworkListener.stop(this)
     }